♻️ refactor : LottoMachine 보완
